split the logic using useSearchParams() into a child component, then wrap that component with <Suspense>
